@@ -1,4 +1,7 @@
-var $ = require('jquery');
-require('jquery-tab');
+import $ from 'jquery';
+import 'jquery-tab';
+import 'jquery-tab/src/built/theme/gray';
+
+import '../css/main.css';
 
 $('.tab1').tab();
