@@ -1,4 +1,7 @@
-import $ = require('jquery');
+import $ from 'jquery';
 import 'jquery-tab';
+import 'jquery-tab/src/built/theme/gray';
+
+import '../css/main.css';
 
 $('.tab1').tab();
