@@ -1,21 +1,12 @@
-const webpack = require('webpack');
 const path = require('path');
 
-module.exports = [{
+module.exports = {
+	mode: 'production',
 	entry: './src/ts/main.ts',
 	output: {
-		libraryTarget: 'umd',
 		path: path.resolve(__dirname, 'dist'),
 		filename: 'bundle.js'
 	},
-	/*externals: {
-		jquery: {
-			commonjs: 'jquery',
-			commonjs2: 'jquery',
-			amd: 'jquery',
-			root: 'jQuery'
-		}
-	},*/
 	resolve: {
 		extensions: ['.ts', '.tsx', '.js', '.jsx']
 	},
@@ -24,53 +15,12 @@ module.exports = [{
 			{
 				test: /\.tsx?$/,
 				loader: 'awesome-typescript-loader'
-			}
-		]
-	},
-	plugins: [
-		new webpack.DefinePlugin({
-			'process.env': {
-				'NODE_ENV': JSON.stringify('production')
-			}
-		})
-	],
-	devtool: 'source-map'
-}, {
-	entry: './src/ts/with-css.ts',
-	output: {
-		libraryTarget: 'umd',
-		path: path.resolve(__dirname, 'dist'),
-		filename: 'bundle-with-css.js'
-	},
-	/*externals: {
-		jquery: {
-			commonjs: 'jquery',
-			commonjs2: 'jquery',
-			amd: 'jquery',
-			root: 'jQuery'
-		}
-	},*/
-	resolve: {
-		extensions: ['.ts', '.tsx', '.js', '.jsx']
-	},
-	module: {
-		rules: [
-			{
-				test: /\.css$/,
-				loader: ['style-loader', 'css-loader']
 			},
 			{
-				test: /\.tsx?$/,
-				loader: 'awesome-typescript-loader'
+				test: /\.css$/,
+				use: ['style-loader', 'css-loader']
 			}
+
 		]
-	},
-	plugins: [
-		new webpack.DefinePlugin({
-			'process.env': {
-				'NODE_ENV': JSON.stringify('production')
-			}
-		})
-	],
-	devtool: 'source-map'
-}];
+	}
+};
